requirements.txt
emoji==2.14.1
fastapi==0.115.11
langdetect==1.0.9
librosa==0.10.2.post1
locust==2.33.1
moviepy==2.1.2
noisereduce==3.0.3
openai==1.66.3
protobuf==6.30.1
pyaudioanalysis==0.3.14
matplotlib>=3.4.2
simplejson>=3.16.0
scipy>=1.6.3
numpy>=1.20.3
hmmlearn>=0.2.5
eyeD3>=0.9.6
pydub>=0.25.1
scikit_learn>=0.24.2
tqdm>=4.52.0
plotly>=5.3.1
pandas>=1.2.4
imblearn
pydantic==2.10.6
python-dotenv==1.0.1
redis==5.2.1
Requests==2.32.3
rq==2.1.0
silero_vad==5.1.2
soundfile==0.13.1